requirements.txt
asgiref
Brotli
certifi
cffi
charset-normalizer
cssselect2
Django
django-admin-interface
django-adminlte-3
django-adminlte-ui
django-colorfield
django-cors-headers
djangorestframework
djangorestframework-simplejwt
docopt
fonttools
html5lib
idna
num2words
pillow
pycparser
pydyf
PyJWT
PyMySQL
pyphen
python-barcode
python-slugify
pytz
qrcode
requests
six
sqlparse
text-unidecode
tinycss2
typing_extensions
urllib3
weasyprint
webencodings
zopfli
gunicorn
